Improvin overall performance. Added ISR encoder, unified variables types. Updated display

+++ lib/MyGlass/GlassConfig.h
@@ -3,7 +3,7 @@
 struct GlassConfig{
   unsigned int dout;              // Beam dout pin
   unsigned int sck;               // Beam sck pin
-  float angle;                    // Servo angle to set
+  int angle;                      // Servo angle to set
 };
 
 #define GLASS_THRESHOLD 10        // Minimal weight to trigger glass
